thêm đ ăng nhập đăng ký

+++ Base/src/pages/List.tsx
@@ -1,13 +1,12 @@
 import { useEffect, useState } from "react";
-import { INotes } from "../interface/Notes";
+import type { INotes } from "../interface/Notes";
 import axios from "axios";
 import toast from "react-hot-toast";
 import { Link } from "react-router-dom";
 
-
 function List() {
-  const API="http://localhost:3000"
-  const[notes,setNotes]=useState<INotes[]>([]);
+  const API="http://localhost:3000";
+  const [notes,setNotes]=useState<INotes[]>([]);
 
   useEffect(()=>{
     const getAll=async()=>{
@@ -15,28 +14,25 @@ function List() {
         const res=await axios.get(`${API}/notes`);
         setNotes(res.data);
         toast.success("Thành công");
-
       }catch(error){
-        toast.success("Không thành công")
+        toast.error("Lỗi ")
       }
     }
     getAll();
   },[])
 
-const Delete=async(id:any)=>{
-  try{
-    await axios.delete(`${API}/notes/${id}`);
-    setNotes(notes.filter(t=> t.id !== id));
-    toast.success("Xóa thành công")
-  }catch(error){
-    toast.error("Xóa k thành công")
+  const Delete=async(id:any)=>{
+    const confirm=window.confirm("Bạn có muốn xóa k ?");
+    if(!confirm) return;
+    try{
+      await axios.delete(`${API}/notes/${id}`);
+      setNotes(notes.filter(t => t.id !== id));
+      toast.success("Xóa thành công")
+    }catch(error){
+      toast.error("Xóa k thành công")
+    }
   }
-}
-
-
-
  
-
   return (
     <div className="p-6">
       <h1 className="text-2xl font-semibold mb-6 text-green-700">Danh sách</h1>
@@ -65,8 +61,9 @@ const Delete=async(id:any)=>{
           </thead>
 
           <tbody>
+
             {notes.map((item,index)=>
-               <tr className="hover:bg-green-50" key={item.id}>
+              <tr className="hover:bg-green-50" key={item.id}>
                 <td className="px-4 py-2 border border-green-300">{index+1}</td>
                 <td className="px-4 py-2 border border-green-300">{item.title}</td>
                 <td className="px-4 py-2 border border-green-300">{item.content}</td>
@@ -79,12 +76,15 @@ const Delete=async(id:any)=>{
                   <button onClick={()=>Delete(`${item.id}`)}>Xóa</button>
                   <Link to={`/notes/edit/${item.id}`}>Sửa</Link>
                 
+                
                 </td>
               </tr>
-
             )}
            
-             
+          
+
+              
+      
           </tbody>
         </table>
       </div>
